requirements.txt
Django>=3.0,<4.0
psycopg2>=2.8
django-crispy-forms>=1.12.0
Pillow
django_taggit==1.2.0
markdown==3.2.1